Remove visit dates
