+++ tests/management/test_management.py
@@ -66,121 +66,6 @@ class ContactModelTests(TestCase):
 		past_contact = Contact(date_of_birth=past_date)
 		self.assertIs(past_contact.has_been_born(), True)
 
-	def test_set_visit_dates_standards_with_today_birth(self):
-		"""
-		set_visit_dates() returns the same standard_dates as manually adding the correct times to
-		today's date for a Cotanct with the date_of_birth of today
-		"""
-		today_contact = create_contact(name="Today", days=0)
-		standards, functionals = today_contact.set_visit_dates()
-
-		today = datetime.date.today()
-		dates = get_scheduled_dates(today)
-
-		self.assertEqual(standards["six_weeks"] == dates["six_weeks"], True)
-		self.assertEqual(standards["ten_weeks"] == dates["ten_weeks"], True)
-		self.assertEqual(standards["fourteen_weeks"] == dates["fourteen_weeks"], True)
-		self.assertEqual(standards["nine_months"] == dates["nine_months"], True)
-		self.assertEqual(standards["sixteen_months"] == dates["sixteen_months"], True)
-		self.assertEqual(standards["five_years"] == dates["five_years"], True)
-
-	def test_set_visit_dates_standards_with_past_birth(self):
-		"""
-		set_visit_dates() returns the same standard_dates as manually adding the correct times to
-		today's date for a Cotanct with the date_of_birth in the past
-		"""
-		num_days_in_past = -10
-		past_contact = create_contact(name="Past", days=num_days_in_past)
-		standards, functionals = past_contact.set_visit_dates()
-
-		past_date = datetime.date.today() + relativedelta(days=num_days_in_past)
-		dates = get_scheduled_dates(past_date)
-
-		self.assertEqual(standards["six_weeks"] == dates["six_weeks"], True)
-		self.assertEqual(standards["ten_weeks"] == dates["ten_weeks"], True)
-		self.assertEqual(standards["fourteen_weeks"] == dates["fourteen_weeks"], True)
-		self.assertEqual(standards["nine_months"] == dates["nine_months"], True)
-		self.assertEqual(standards["sixteen_months"] == dates["sixteen_months"], True)
-		self.assertEqual(standards["five_years"] == dates["five_years"], True)
-
-	def test_set_visit_dates_standards_with_future_birth(self):
-		"""
-		set_visit_dates() returns the same standard_dates as manually adding the correct times to
-		today's date for a Cotanct with the date_of_birth in the future
-		"""
-		num_days_in_future = 400
-		future_contact = create_contact(name="Future", days=num_days_in_future)
-		standards, functionals = future_contact.set_visit_dates()
-
-		future_date = datetime.date.today() + relativedelta(days=num_days_in_future)
-		dates = get_scheduled_dates(future_date)
-
-		self.assertEqual(standards["six_weeks"] == dates["six_weeks"], True)
-		self.assertEqual(standards["ten_weeks"] == dates["ten_weeks"], True)
-		self.assertEqual(standards["fourteen_weeks"] == dates["fourteen_weeks"], True)
-		self.assertEqual(standards["nine_months"] == dates["nine_months"], True)
-		self.assertEqual(standards["sixteen_months"] == dates["sixteen_months"], True)
-		self.assertEqual(standards["five_years"] == dates["five_years"], True)
-
-
-	def test_set_visit_dates_functionals_with_today_birth(self):
-		"""
-		set_visit_dates() returns the same functional_dates as manually adding the correct times to
-		today's date for a Cotanct with the date_of_birth of today
-		"""
-		today_contact = create_contact(name="Today", days=0)
-		standards, functionals = today_contact.set_visit_dates()
-
-		today = datetime.date.today()
-		dates = get_scheduled_dates(today)
-
-		self.assertEqual(functionals["six_weeks"] == dates["six_weeks"], True)
-		self.assertEqual(functionals["ten_weeks"] == dates["ten_weeks"], True)
-		self.assertEqual(functionals["fourteen_weeks"] == dates["fourteen_weeks"], True)
-		self.assertEqual(functionals["nine_months"] == dates["nine_months"], True)
-		self.assertEqual(functionals["sixteen_months"] == dates["sixteen_months"], True)
-		self.assertEqual(functionals["five_years"] == dates["five_years"], True)
-
-	def test_set_visit_dates_functionals_with_past_birth(self):
-		"""
-		set_visit_dates() returns the same functional_dates as manually adding the correct times to
-		today's date for a Cotanct with the date_of_birth in the past
-		"""
-		num_days_in_past = -10
-		past_contact = create_contact(name="Past", days=num_days_in_past)
-		past_date = datetime.date.today() + relativedelta(days=num_days_in_past)
-		past_contact.functional_date_of_birth = past_date
-
-		standards, functionals = past_contact.set_visit_dates()
-		dates = get_scheduled_dates(past_date)
-
-		self.assertEqual(functionals["six_weeks"] == dates["six_weeks"], True)
-		self.assertEqual(functionals["ten_weeks"] == dates["ten_weeks"], True)
-		self.assertEqual(functionals["fourteen_weeks"] == dates["fourteen_weeks"], True)
-		self.assertEqual(functionals["nine_months"] == dates["nine_months"], True)
-		self.assertEqual(functionals["sixteen_months"] == dates["sixteen_months"], True)
-		self.assertEqual(functionals["five_years"] == dates["five_years"], True)
-
-	def test_set_visit_dates_functionals_with_future_birth(self):
-		"""
-		set_visit_dates() returns the same functional_dates as manually adding the correct times to
-		today's date for a Cotanct with the date_of_birth in the future
-		"""
-		num_days_in_future = 400
-		future_contact = create_contact(name="Future", days=num_days_in_future)
-		future_date = datetime.date.today() + relativedelta(days=num_days_in_future)
-		future_contact.functional_date_of_birth = future_date
-
-		standards, functionals = future_contact.set_visit_dates()
-		dates = get_scheduled_dates(future_date)
-
-		self.assertEqual(standards["six_weeks"], dates["six_weeks"])
-		self.assertEqual(standards["ten_weeks"], dates["ten_weeks"])
-		self.assertEqual(standards["fourteen_weeks"], dates["fourteen_weeks"])
-		self.assertEqual(standards["nine_months"], dates["nine_months"])
-		self.assertEqual(standards["sixteen_months"], dates["sixteen_months"])
-		self.assertEqual(standards["five_years"], dates["five_years"])
-
 
 class ContactIndexViewTests(TestCase):
     def test_no_contacts(self):
